+++ src/components/genre-list/genre-list.test.js
@@ -1,27 +1,15 @@
 import React from "react";
 import renderer from "react-test-renderer";
 import {GenreList} from "./genre-list.jsx";
-import {Provider} from "react-redux";
-import configureStore from "redux-mock-store";
 import {films} from "../../test-data";
 
-const mockStore = configureStore([]);
-
-it(`Render GenreList`, () => {
-  const store = mockStore({
-    movies: films,
-    activeGenre: `All genres`,
-  });
-
+it(`GenreList renders correctly`, () => {
   const tree = renderer
     .create(
-        <Provider store={store}>
-          <GenreList
-            movies={films}
-            activeGenre={`All genres`}
-            changeGenre={() => {}}
-          />
-        </Provider>,
+        <GenreList
+          activeGenre={`All genres`}
+          movies={films}
+          changeGenre={() => {}} />,
         {
           createNodeMock: () => {
             return {};
@@ -29,6 +17,5 @@ it(`Render GenreList`, () => {
         }
     )
     .toJSON();
-
   expect(tree).toMatchSnapshot();
 });
